feat: match existing service files in style and service call, merging into one notebook
